+++ src/vscode/treeviews/Workspaces/FabricItem.ts
@@ -16,6 +16,7 @@ import { FabricFSUri } from '../../filesystemProvider/FabricFSUri';
 import { FabricConfiguration } from '../../configuration/FabricConfiguration';
 import { FabricApiTreeItem } from '../FabricApiTreeItem';
 import { FabricApiService } from '../../../fabric/FabricApiService';
+import { FabricItemTags } from './FabricItemTags';
 
 // https://vshaxe.github.io/vscode-extern/vscode/TreeItem.html
 export class FabricItem extends FabricWorkspaceTreeItem {
@@ -88,6 +89,19 @@ export class FabricItem extends FabricWorkspaceTreeItem {
 			if (this.contextValue.includes("EDIT_")) {
 				children.push(new FabricItemDefinition(this));
 			}
+
+			try {
+				let tags = new FabricItemTags(this);
+				const tagChildren = await FabricApiTreeItem.getValidChildren(tags) as FabricWorkspaceTreeItem[];
+				if (tagChildren.length > 0) {
+					tagChildren.forEach(child => tags.addChild(child));
+					children.push(tags);
+				}
+			}
+			catch (e) {
+				ThisExtension.Logger.logInfo("Could not load tags for item " + this.itemName);
+			}
+
 			// Connections
 			let supportedItemTypes: FabricApiItemType[] = [];
 			try {